% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{DataProcessing}
\alias{DataProcessing}
\title{DataProcessing}
\usage{
DataProcessing(Data = NULL, Method = "Standard")
}
\arguments{
\item{Data}{a numeric data.frame.}

\item{Method}{character indicating the method used to Data preprocessing. If data are continous, use "Standard".
If Data are compositional, please use "Compositional" and clr (centered log-ratios functions) transformations are performed.
To compositional data, you also could use the option "TSS" Total Sum Scaling follow up bray (Bray-Curtis) in distance option.
The function also allows to processing frequencies- like data through "FreqNorm" option.
Note that when you use Compositional, we first sum 1 to all the counts (in order to performs the log transformation before).}
}
\value{
a data.frame with normalized data.
}
\description{
function to Perform external datas' pre-processing.
This function allows an external pre-processing of the datasets including on the analysis in three ways: Standard, Compositional (centered log ratio) and
frequencies.
}
\examples{
{
data(Taraoceans)
Data<-Taraoceans$phychem
Data<-DataProcessing(Data,Method="Standard")
}
}
\author{
Laura M Zingatetti
}
