% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistStatis-Class.R
\docType{methods}
\name{DistStatis-class}
\alias{DistStatis-class}
\title{Class \code{DistStatis} DistStatis S4 class (linkHD:Multiple Heterogeneous Dataset Integration)
Statis with Distance options implementation.}
\description{
Class \code{DistStatis} DistStatis S4 class (linkHD:Multiple Heterogeneous Dataset Integration)
Statis with Distance options implementation.
}
\section{Features}{

\enumerate{
  \item DistStatis (implements Statis method incorporating Distance options to integrate multiple heterogeneous datasets)
  \item Implement a LM (Linear Model) to variable selection
  \item Incorporate a method to variable clustering
  \item Incorporate some visualization tools: Compromise visualization, Relationship-visualization
}
}

\section{Fields}{

\itemize{
 \item RV: Vectorial Correlation Matrix between studies.
 \item Inertia.RV: Inertia (\%) explained for all tables.
 \item Euclid.Im: Euclidean Image of all studies.
 \item Inertia.comp: Inertia (\%) explained for all dimensions of compromise matrix.
 \item Compromise.Coords: Projection of all observations in compromise (Coords).
 \item Compromise.Matrix: Compromise Matrix from statis methodology.
 \item RQO: Representation Quality of observations in compromise matrix.
 \item TableProjections: Projection of each table on Compromise configuration
}

 @slot RV: Vectorial Correlation Matrix between studies.
 @slot  Inertia.RV: Inertia (\%) explained for all tables.
 @slot  Euclid.Im: Euclidean Image of all studies.
 @slot  Inertia.comp: Inertia (\%) explained for all dimensions of compromise matrix.
 @slot  Compromise.Coords: Projection of all observations in compromise (Coords).
 @slot  Compromise.Matrix: Compromise Matrix from statis methodology.
 @slot  RQO: Representation Quality of observations in compromise matrix.
 @slot  TableProjections: Projection of each table on Compromise configuration
}

\section{DistStatis-general-functions}{

\describe{
 \item{DistStatis}{Getters for their respective slots.}
}

 @author Laura M Zingaretti
}

\examples{
{
showClass('DistStatis')
}
}
