% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing.R
\name{impute_missing}
\alias{impute_missing}
\title{impute_missing}
\usage{
impute_missing(data)
}
\arguments{
\item{data}{Data frame with abundances values. Note that this function should only be used with the raw data (counts)}
}
\value{
Imputed data frame
}
\description{
This function imputes missing abundances by sampling the known values from each variable.
The values are sampled according to the distribution, i.e. the most frequent values have a better chance of being chosen.
}
\examples{
{
# toy example. To simulate 13 missing columns with less than 50 \\% of missing values in each.

data('Ruminotypes')
Data<-Ruminotypes$`16_S`
#13 indicates the number of columns with missing values.
Columns<-sample(1:ncol(Data),13)
for (i in Columns){
n<-sample(1:30,1)
Data[sample(1:nrow(Data),n),i]<-NA
}

A<-impute_missing(Data)
#check precision of imputed data
cor(A[,Columns[1]],Ruminotypes$`16_S`[,Columns[1]])
cor(A[,Columns[2]],Ruminotypes$`16_S`[,Columns[2]])
}



}
\author{
Laura M Zingatetti
}
