% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_inspect.R
\name{NA_inspect}
\alias{NA_inspect}
\title{NA_inspect}
\usage{
NA_inspect(dat)
}
\arguments{
\item{dat}{list with your datasets. It can be the same list generated by ReadData function.}
}
\value{
List with missing values information or message if your data doesn't have missing
}
\description{
This function inspects your datasets (stored int a list)
in order to evaluate if your data contains na values.
}
\examples{
{
data('Ruminotypes')
Data<-Ruminotypes$`16_S`
#13 indicates the number of columns with missing values.
Columns<-sample(1:ncol(Data),13)
for (i in Columns){
n<-sample(1:30,1)
Data[sample(1:nrow(Data),n),i]<-NA
}

Data_1<-list("16_S"=Ruminotypes$`16_S`,"16_S_NA"=Data)
NA_inspect(Data_1)
}

}
\author{
Laura M Zingatetti
}
