% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Data.R
\name{Read_Data}
\alias{Read_Data}
\title{Read_Data: a fast way to data reading.}
\usage{
Read_Data(Path = "")
}
\arguments{
\item{Path}{path to folder containing all dataset to integrate}
}
\value{
\item{List}{List including all dataset into the parent directory. List names inherit the  names of the files}
}
\description{
this function read all dataset in a folder and returns list needed to Link_Data function input.
}
\examples{

###Datos<-Read_Data("Path to parent folder",common_elements=1)


}
\author{
Laura M Zingatetti
}
